% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{declared_internals}
\alias{declared_internals}
\alias{example_internal}
\alias{format_declared}
\alias{order_declared}
\alias{value_labels}
\alias{variable_label}
\alias{check_measurement}
\alias{all_missing_values}
\alias{names_values}
\alias{stopError_}
\alias{coerceMode_}
\alias{possibleNumeric_}
\alias{asNumeric_}
\alias{wholeNumeric_}
\alias{tryCatchWEM_}
\alias{padLeft_}
\alias{padRight_}
\alias{padBoth_}
\alias{unlockEnvironment_}
\alias{hasTag_}
\alias{getTag_}
\alias{numdec_}
\alias{trimstr_}
\title{declared internal functions}
\usage{
example_internal()

format_declared(x, digits = getOption("digits"))

order_declared(
  x,
  na.last = NA,
  decreasing = FALSE,
  method = c("auto", "shell", "radix"),
  empty.last = na.last
)

value_labels(...)

variable_label(...)

check_measurement(x)

all_missing_values(x, na_values = NULL, na_range = NULL, labels = NULL)

names_values(x, drop_na = FALSE)

stopError_(message, enter = "\\n")

coerceMode_(x)

possibleNumeric_(x, each = FALSE)

asNumeric_(x, levels = TRUE)

wholeNumeric_(x, each = FALSE)

tryCatchWEM_(expr, capture = FALSE)

padLeft_(x, n)

padRight_(x, n)

padBoth_(x, n)

unlockEnvironment_(env)

hasTag_(x, tag = NULL)

getTag_(x)

numdec_(x, each = FALSE, na.rm = TRUE, maxdec = 15)

trimstr_(x, what = " ", side = "both")
}
\description{
Only to be used internally for developers and contributors.
}
\details{
The following functions are copied from package admisc to achieve zero dependency
}
\section{admisc functions}{

\code{\link{stopError_}}, \code{\link{coerceMode_}},
}

\keyword{internal}
